move test_timestamps-->test_timeseries per reviewer request
